Crate buck2_test_proto

Modules

arg_value_content

Nested message and enum types in ArgValueContent.

execute_response2

Nested message and enum types in ExecuteResponse2.

execution_status

Nested message and enum types in ExecutionStatus.

execution_stream

Nested message and enum types in ExecutionStream.

external_runner_spec_value

Nested message and enum types in ExternalRunnerSpecValue.

output

Nested message and enum types in Output.

remote_object

Nested message and enum types in RemoteObject.

test_executor_client

Generated client implementations.

test_executor_server

Generated server implementations.

test_orchestrator_client

Generated client implementations.

test_orchestrator_server

Generated server implementations.

test_result

Nested message and enum types in TestResult.

test_stage

Nested message and enum types in TestStage.

Structs

ArgFormat

ArgValue

ArgValueContent

AttachInfoMessageRequest

Cancelled

CasDigest

Copy of Digest from Remote Execution API: https://github.com/facebook/buck2/blob/main/remote_execution/oss/re_grpc_proto/proto/build/bazel/remote/execution/v2/remote_execution.proto#L936-L943

ConfiguredTarget

ConfiguredTargetHandle

DeclaredOutput

Empty

EndOfTestResultsRequest

EnvironmentVariable

ExecuteRequest2

ExecuteResponse2

ExecutionDetails

ExecutionResult2

ExecutionStatus

ExecutionStream

ExecutorConfigOverride

ExternalRunnerSpec

A spec representing test targets aiming for maximum compatibility with the external runner spec as defined in buck1.

ExternalRunnerSpecRequest

ExternalRunnerSpecValue

LocalResourceType

Output

OutputEntry

OutputName

This message should be compatible with the prefix of DeclaredOutput for backwards compatibility

PrepareForLocalExecutionRequest

PrepareForLocalExecutionResponse

PrepareForLocalExecutionResult

RemoteDirNode

RemoteFileNode

TODO(arr): consider using remote_execution.proto later on. The API surface there is larger but consolidation may be worth it.

RemoteObject

ReportTestResultRequest

ReportTestSessionRequest

ReportTestsDiscoveredRequest

SetupLocalResourceLocalExecutionCommand

TestExecutable

TestResult

TestStage

Testing

Message sent when a TestSuite is discovered.

TtlConfig

UnstableHeapDumpRequest

UnstableHeapDumpResponse

VerbatimEnvironmentVariable

Enums

CancellationReason

TestStatus